requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
channels==4.3.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
PyJWT==2.10.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
